requirements.txt
Pillow==8.0.1
tensorflow==2.3.1
pandas==1.1.4
